Added Unix Makefile + GitHub action to build and test
The Makefile builds the library and the Ym2Wav program. Testing runs all
the sample YM files through Ym2Wav (which will fail with a non-zero exit
code if something goes wrong, failing the build).

I had to rename all sample files with lower case `.ym` extension and
remove the spaces in filenames for this to work properly.

Added GitHub action to build and run the tests.